+++ internal/provider/data_source_iosxe_system_test.go
@@ -17,7 +17,6 @@ func TestAccDataSourceIosxeSystem(t *testing.T) {
 				Config: testAccDataSourceIosxeSystemPrerequisitesConfig + testAccDataSourceIosxeSystemConfig,
 				Check: resource.ComposeTestCheckFunc(
 					resource.TestCheckResourceAttr("data.iosxe_system.test", "hostname", "ROUTER-1"),
-					resource.TestCheckResourceAttr("data.iosxe_system.test", "ip_routing", "true"),
 					resource.TestCheckResourceAttr("data.iosxe_system.test", "ipv6_unicast_routing", "true"),
 					resource.TestCheckResourceAttr("data.iosxe_system.test", "multicast_routing", "true"),
 					resource.TestCheckResourceAttr("data.iosxe_system.test", "multicast_routing_distributed", "true"),
@@ -53,7 +52,6 @@ const testAccDataSourceIosxeSystemConfig = `
 
 resource "iosxe_system" "test" {
   hostname = "ROUTER-1"
-  ip_routing = true
   ipv6_unicast_routing = true
   multicast_routing = true
   multicast_routing_distributed = true